requirements.txt
redis
pycryptodome
pymysql
sqlalchemy
func-timeout
websocket-client
requests